requirements.txt
asgiref==3.8.1
Django==5.2
django-bootstrap5==25.1
django-ckeditor5==0.0.4
django-js-asset==3.1.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1

django-filter~=25.1